Update main.py
I've implemented codes to prompt the user to begin/restart the program. Other changes were made (mainly rewording the comments). The codes on line 16 through 28 and 98 through 113 are not completely finished. The program runs fine, but I'm unable to make the second half of those line for codes to work.